--- integration-tests/public/runtime-call-contract/e2e_tests/mod.rs
@@ -1,26 +1,0 @@
-mod runtime;
-
-use super::flipper::*;
-use ink_e2e::{
-    subxt::dynamic::Value,
-    ChainBackend,
-    ContractsBackend,
-};
-
-type E2EResult<T> = std::result::Result<T, Box<dyn std::error::Error>>;
-
-/// Just instantiate a contract using non-default runtime.
-#[ink_e2e::test(backend(runtime_only(sandbox = runtime::ContractCallerSandbox)))]
-async fn custom_runtime<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
-    client
-        .instantiate(
-            "runtime-call-contract",
-            &ink_e2e::alice(),
-            &mut FlipperRef::new(false),
-        )
-        .submit()
-        .await
-        .expect("instantiate failed");
-
-    Ok(())
-}
