+++ integration-tests/public/runtime-call-contract/flipper/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "flipper"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+ink = { workspace = true }
+flipper-traits = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+    "ink/std",
+]
+ink-as-dependency = []
