+++ integration-tests/public/runtime-call-contract/lib.rs
@@ -1,44 +1,23 @@
-#![cfg_attr(not(feature = "std"), no_std, no_main)]
+use flipper::*;
+use ink_e2e::{
+    ChainBackend,
+    ContractsBackend,
+};
 
-#[ink::contract]
-pub mod flipper {
-    #[ink(storage)]
-    pub struct Flipper {
-        value: bool,
-    }
+type E2EResult<T> = std::result::Result<T, Box<dyn std::error::Error>>;
 
-    impl Flipper {
-        /// Creates a new flipper smart contract initialized with the given value.
-        #[ink(constructor)]
-        pub fn new(init_value: bool) -> Self {
-            Self { value: init_value }
-        }
+/// Just instantiate a contract using non-default runtime.
+#[ink_e2e::test(backend(runtime_only(sandbox = sandbox_runtime::ContractCallerSandbox)))]
+async fn custom_runtime<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
+    client
+        .instantiate(
+            "runtime-call-contract",
+            &ink_e2e::alice(),
+            &mut FlipperRef::new(false),
+        )
+        .submit()
+        .await
+        .expect("instantiate failed");
 
-        /// Creates a new flipper smart contract initialized to `false`.
-        #[ink(constructor)]
-        pub fn new_default() -> Self {
-            Self::new(Default::default())
-        }
-
-        /// Flips the current value of the Flipper's boolean.
-        #[ink(message)]
-        pub fn flip(&mut self) {
-            self.value = !self.value;
-        }
-
-        /// Returns the current value of the Flipper's boolean.
-        #[ink(message)]
-        pub fn get(&self) -> bool {
-            self.value
-        }
-
-        /// Returns the current balance of the Flipper.
-        #[ink(message)]
-        pub fn get_contract_balance(&self) -> Balance {
-            self.env().balance()
-        }
-    }
+    Ok(())
 }
-
-#[cfg(all(test, feature = "e2e-tests"))]
-mod e2e_tests;
