+++ integration-tests/public/runtime-call-contract/sandbox-runtime/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "sandbox-runtime"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+pallet-contract-caller = { path = "pallet-contract-caller", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+    "pallet-contract-caller/std",
+]
