+++ integration-tests/public/runtime-call-contract/sandbox-runtime/src/lib.rs
@@ -5,10 +5,10 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 ink_sandbox::create_sandbox!(ContractCallerSandbox, (), (), {
-    ContractCaller: super::pallet_contract_caller,
+    ContractCaller: pallet_contract_caller,
 });
 
 impl pallet_contract_caller::Config
-    for <ContractCallerSandbox as ink_sandbox::Sandbox>::Runtime
+for <ContractCallerSandbox as ink_sandbox::Sandbox>::Runtime
 {
 }
